requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
autocorrect==2.5.0
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.1.5
docopt==0.6.2
filelock==3.0.12
flatbuffers==1.12
fontawesome-free==5.15.3
gast==0.3.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
inexactsearch==1.0.2
Js2Py==0.70
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mysqlclient==2.0.3
natsort==7.1.1
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.7
pandas==1.2.3
pbr==5.5.1
Pillow==8.1.0
pipwin==0.5.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
PyQt5==5.15.1
pyqt5-plugins==5.15.1.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.3
PyQt5Designer==5.14.1
pySmartDL==1.3.4
pyspellchecker==0.6.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
qt5-applications==5.15.1.2.1
qt5-tools==5.15.1.1.0.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
silpa-common==0.3
six==1.15.0
soundex==1.1.3
soupsieve==2.0.1
spellchecker==0.4
sqlparse==0.4.1
stevedore==3.3.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
virtualenv==16.7.9
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
wrapt==1.12.1
wxPython==4.1.1
